- Remove full_name from SelectionCommon and related code - custom enum tests pass
