--- rust/shalom_dart_codegen/tests/usecases/mod.rs
@@ -1,10 +1,0 @@
-mod test_custom_scalar;
-mod test_custom_scalar_arguments;
-mod test_enum_arguments;
-mod test_enum_selection;
-mod test_input_objects;
-mod test_list_scalars;
-mod test_nested_input_objects;
-mod test_object_selection;
-mod test_operation_scalar_arguments;
-mod test_simple_scalars;
