--- rust/shalom_dart_codegen/tests/usecases/test_enum_selection.rs
@@ -1,8 +1,0 @@
-use crate::run_dart_tests_for_usecase;
-
-static USE_CASE_NAME: &str = "enum_selection";
-
-#[test]
-fn test_enum_selection_dart() {
-    run_dart_tests_for_usecase(USE_CASE_NAME);
-}
