--- rust/shalom_dart_codegen/tests/usecases/test_object_selection.rs
@@ -1,12 +1,0 @@
-use crate::run_dart_tests_for_usecase;
-
-static USE_CASE_NAME: &str = "object_selection";
-#[test]
-fn test_object_selection_dart() {
-    run_dart_tests_for_usecase(USE_CASE_NAME);
-}
-
-#[test]
-fn test_nested_object_selection_dart() {
-    run_dart_tests_for_usecase("nested_object_selection");
-}
